refactor: move violation methods in a new file
